+++ tox.ini
@@ -246,7 +246,6 @@ deps =
     framework_falcon-falcon0200: falcon<2.1
     framework_falcon-falconmaster: https://github.com/falconry/falcon/archive/master.zip
     framework_fastapi: fastapi
-    framework_fastapi: graphene
     framework_fastapi: asyncio
     framework_flask: Flask-Compress
     framework_flask-flask0012: flask<0.13
@@ -281,6 +280,7 @@ deps =
     framework_sanic-sanic210300: sanic<21.3.1
     framework_sanic-saniclatest: sanic
     framework_sanic-sanic{1812,190301,1906}: aiohttp
+    framework_starlette: graphene
     framework_starlette-starlette0014: starlette<0.15
     framework_starlette-starlettelatest: starlette
     framework_strawberry: starlette
